Build the docs folder
