+++ docs/main.js
@@ -267,7 +267,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _displayLikes_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./displayLikes.js */ \"./src/modules/displayLikes.js\");\n/* harmony import */ var _addLikes_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./addLikes.js */ \"./src/modules/addLikes.js\");\n/* harmony import */ var _pop_up_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./pop-up.js */ \"./src/modules/pop-up.js\");\n\n\n\n\nconst list = document.querySelector('#list');\n\nconst populate = async (movies) => {\n  const likeArray = await (0,_displayLikes_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])();\n  movies.forEach((movie, index) => {\n    if (index < 24) {\n      let noOfLikes = 0;\n      likeArray.forEach((like) => {\n        if (like.item_id === movie.imdbid) {\n          noOfLikes = like.likes;\n        }\n      });\n      list.innerHTML\n    += `\n    <div class=\"grid-item\">\n    <figure>\n      <img src=\"${movie.imageurl[0]}\" alt=\"Cover art of movie\">\n    </figure>\n\n    <h2 class=\"movie-title\">${movie.title}</h2>\n\n    <div class=\"my-flex\">\n      <button type=\"button\" class=\"btn my-btn\" id=\"${movie.imdbid}\">Comments</button>\n      <button type=\"button\" class=\"btn my-btn1\" id= \"like${index}\">\n        <span>${noOfLikes}</span>\n        <i class=\"bi bi-heart\"></i>\n      </button>\n    </div>\n  </div>\n    `;\n    }\n  });\n\n  const likebtn = Array.from(document.querySelectorAll('.my-btn1'));\n\n  await likebtn.forEach(async (btn) => {\n    const btnID = btn.id;\n    await movies.forEach(async (elem, index) => {\n      const iD = `like${index}`;\n      if (btnID === iD) {\n        await btn.addEventListener('click', async () => {\n          (0,_addLikes_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"])(elem.imdbid);\n          const likeArr = await (0,_displayLikes_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])();\n          likeArr.forEach((like) => {\n            if (like.item_id === elem.imdbid) {\n              btn.querySelector('span').innerHTML = like.likes;\n            }\n          });\n        });\n      }\n    });\n  });\n  (0,_pop_up_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(movies);\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (populate);\n\n//# sourceURL=webpack://MediaHub/./src/modules/createList.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _displayLikes_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./displayLikes.js */ \"./src/modules/displayLikes.js\");\n/* harmony import */ var _addLikes_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./addLikes.js */ \"./src/modules/addLikes.js\");\n/* harmony import */ var _pop_up_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./pop-up.js */ \"./src/modules/pop-up.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n\n\n\n\n\nconst list = document.querySelector('#list');\n\nconst populate = async (movies) => {\n  const likeArray = await (0,_displayLikes_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])();\n  movies.forEach((movie, index) => {\n    if (index < 24) {\n      let noOfLikes = 0;\n      likeArray.forEach((like) => {\n        if (like.item_id === movie.imdbid) {\n          noOfLikes = like.likes;\n        }\n      });\n      list.innerHTML\n    += `\n    <div class=\"grid-item\">\n    <figure>\n      <img src=\"${movie.imageurl[0]}\" alt=\"Cover art of movie\">\n    </figure>\n\n    <h2 class=\"movie-title\">${movie.title}</h2>\n\n    <div class=\"my-flex\">\n      <button type=\"button\" class=\"btn my-btn\" id=\"${movie.imdbid}\">Comments</button>\n      <button type=\"button\" class=\"btn my-btn1\" id= \"like${index}\">\n        <span>${noOfLikes}</span>\n        <i class=\"bi bi-heart\"></i>\n      </button>\n    </div>\n  </div>\n    `;\n    }\n  });\n\n  await (0,_like_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"])(movies);\n  (0,_pop_up_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(movies);\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (populate);\n\n//# sourceURL=webpack://MediaHub/./src/modules/createList.js?");
 
 /***/ }),
 
@@ -301,6 +301,17 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst items = () => {\n  const list = document.querySelector('#list');\n  const movieList = Array.from(list.querySelectorAll('.grid-item'));\n  document.getElementById('num').innerHTML = movieList.length;\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (items);\n\n//# sourceURL=webpack://MediaHub/./src/modules/itemCounter.js?");
+
+/***/ }),
+
+/***/ "./src/modules/like.js":
+/*!*****************************!*\
+  !*** ./src/modules/like.js ***!
+  \*****************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _addLikes_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./addLikes.js */ \"./src/modules/addLikes.js\");\n/* harmony import */ var _displayLikes_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./displayLikes.js */ \"./src/modules/displayLikes.js\");\n\r\n\r\n\r\nconst like = async (movies) => {\r\n  const likebtn = Array.from(document.querySelectorAll('.my-btn1'));\r\n\r\n  await likebtn.forEach(async (btn) => {\r\n    const btnID = btn.id;\r\n    await movies.forEach(async (elem, index) => {\r\n      const iD = `like${index}`;\r\n      if (btnID === iD) {\r\n        await btn.addEventListener('click', async () => {\r\n          await (0,_addLikes_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(elem.imdbid);\r\n          const likeArr = await (0,_displayLikes_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"])();\r\n          likeArr.forEach((like) => {\r\n            if (like.item_id === elem.imdbid) {\r\n              btn.querySelector('span').innerHTML = like.likes;\r\n            }\r\n          });\r\n        });\r\n      }\r\n    });\r\n  });\r\n}\r\n\r\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (like);\n\n//# sourceURL=webpack://MediaHub/./src/modules/like.js?");
 
 /***/ }),
